chemical_symbol_accepts double Z and uses and asterik* when Z != round(Z)

+++ src/chemical_symbol.hxx
@@ -7,7 +7,7 @@
 namespace chemical_symbol {
 
   int8_t decode(char const S, char const y);
-  status_t get(char* Sy, int const Z, char const blank=0); // blank=' ' makes all symbols have a length of 2 chars
+  status_t get(char* Sy, double const Z, char const blank=0); // blank=' ' makes all symbols have a length of 2 chars
 
   status_t all_tests(int const echo=0);
